M2-PR-06: split dbus_session_tests.rs (was 1631 LOC, over the ceiling)
Behaviour-preserving. 15 tests moved from dbus_session_tests.rs into
4 leaf files by what each test exercises:

- dbus_session_status.rs (674, 5): real-bus + initial-layer +
  focus-source + paused-changed signal + status-changed focus signal.
- dbus_session_restart.rs (238, 3): dbus-restart + control-command
  restart + control-command error-without-service.
- dbus_session_pause.rs (627, 6): pause/unpause + focus-ignored
  paused + pause-wayland-env + unfocus-ignored + release-VKs/reset
  + control-command pause/unpause.
- dbus_session_persistent.rs (95, 1): persistent-service handles
  restart in idle.

All four files under 1500 LOC. dbus_session_tests.rs deleted.
integration_tests/mod.rs declarations updated.

Plan inventory delta: M2-PR-05 reported 14 tests in this section;
actual count was 15 (plan §2 missed test_dbus_paused_changed_signal).
The aggregate 65 integration-test count is unchanged.

Verification: 15 test attrs preserved (sum across leaves); cargo
build clean; cargo test --bin kanata-switcher 261 passed / 0 failed.

M2 milestone complete:
- tests.rs (4864 LOC, 189 tests) -> 22 leaf files (max 770 LOC)
- integration_tests.rs (7176 LOC, 65 tests) -> 16 leaf files (max 1463)
- No production source changes; behaviour preserved.
