+++ src/daemon/integration_tests/mod.rs
@@ -27,6 +27,9 @@ mod wayland;
 mod x11;
 mod vk_validation;
 mod dbus_control;
-mod dbus_session_tests;
+mod dbus_session_status;
+mod dbus_session_restart;
+mod dbus_session_pause;
+mod dbus_session_persistent;
 mod dbus_multiplex;
 mod dconf;
